fix : saperate logic of login screen.
